Add search functionality to search saved submisisons #3.
Add a search menu item to SubmissionsFragment's menu to display an
alert dialog. The dialog is used to search downloaded submissions.

Add searchSubmissions to SubmissionsPresenter to search the list of
downloaded submissions.
Add subredditMatched to remove duplication with how a submisison was
matched with a subreddit.

Made public downloadDbSubmissions and downloadNetworkSubmisisons
private, since they did not need to be public (in DataManager).