Add Tests for SubmissionsPresenter
SubmissionsPresenterTest contains tests for some of the
SubmissionsPresenter's functions.

Add Mockito to the gradle file since it is used for the tests.
Add MockSubmission because Mockito can't mock final classes and
Submission is a final class.
Modify SubmissionsPresenter and BasePresenter to accept a DataManager as
a parameter in their constructors since it makes it easier to test the
presenters.
Add RxSchedulesOverrideRule to handle issues with using unit tests and
RxJava (taken from https://goo.gl/yg7avt).

Remove ExampleUnitTest since it was not used.